Reworked delegateBySig to be consistent with permit
Applied some changes to delegateBySig to make it consistent with permit.

First, renamed `expiry` in Delegation struct to `deadline` to keep it
consistent with `Permit`'s `deadline` serving the same purpose.

Next, removed from the parameter list `nonce` and added there
`signatory`. This way, `delegateBySig` has an interface consistent with
`permit` where `owner` is passed as a parameter an no nonce parameter is
passed.

Last but not least, added validation of `v` signature parameter.