T extends MisfundRecovery, updated @thesis/solidity-contracts dependency
The new @thesis/solidity-contracts version introduces MisfundRecovery
contract allowing the contract owner to recover misfunded ERC721 and ERC20.

Additionally, in ERC20WithPermit, it prevents sending tokens to the
token contract address which should prevent most of the misfund cases
we've seen so far on mainnet for KEEP.

Last but not least, _beforeTokenTransfer in ERC20WithPermit has been
renamed to beforeTokenTransfer.

I have updated T token to extend MisfundRecovery. This way, DAO which
will be the owner of T token will be able to recover misfunds in the
future.