Cleared up slither warning in Checkpoint contract
dead-code detector on maxSupply and delegate is not correct - these
functions are used in the inheriting code and may be overwritten.

Found and reported two bugs in the detectors in moveVotingPower function.

incorrect-equality detector in writeCheckpoint is wrong as well - this line
needs to look like that because we want to update the value if we are still
in the context of the given block.