Updated dependency to `@thesis/solidity-contracts` and fixed test
There was a change of naming in the `@thesis/solidity-contracts` from
`nonces` to `nonce`. We need to update the dependency to that project.
We're also updating the name of the T function in `T.test.js` from
`nonces` to `nonce` (the name of the function was updated in the
contract in one of the previous commits).