Pass fromAmount and toAmount in authorization IApplication callbacks
Instead of passing just `toAmount` (previously named `amount`) we now
pass two values. This is helpful for applications that need to update
internal state. One example is an update of a sortition pool for
operator deauthorizing to 0: we want to remove them from the sortition
pool by setting their in-pool weight to 0 while still keeping them
earning rewards using their previous in-pool weight for the entire
duration of authorization decrease delay.