Update rules for NuCypherToken deployment
To be consistent with other scripts we want to cover case of network
forking with hardhat, the network starts as `hardhat` network. In such
case we expect the artifact to be resolved correctly. For a hardhat
network without forking we will deploy stub contract, in any other case
we need a real contract deployed.