Write startStakingTimestamp in stakeKeep function
We do not use this timestamp for validation when keep staker is
undelegating because for KEEP stakers it does not make sense.

We could write this value though when `stakeKeep` is called so that
we do not have to scan `Staked` events to determine if and when
someone staked.

Also, having the same rules for liquid T, NU and KEEP should be
easier to follow and reduce the possibility of bugs in the code
using `getStartStakingTimestamp` function.