Incorporate workflow in the CI flow
We're adding `Solidity` workflow to he CI flow. After npm package gets
published, `ci` projects will be informed about execution of the
workflow and will start execution of the modules which are configured as
direct downstream modules in `keep-network/ci/config/config.json/`.
Changes required on the `ci` repo side to make that work will be
delivered in a separate PR/commit.