Remove artifacts created during install before etherscan-verify
Some time ago a postinstall script was added to the `package.json` as a
workaround for a problem with hardhat not handling situation when two
different contracts with the same name are processed. As a by-product of
this workaround, during yarn install
a `external/npm/@keep-network/keep-core/artifacts` directory is created,
which contains `keep-core` contracts. We don't want to verify those
contracts during execution of `contracts-etherscan-verification` job. To
exclude them, we need to remove those files before `hardhat
etherscan-verify` command is run. The `./external/mainnet` folder
contains files which do not break etherscan verification, so we're
limiting removal of files to `./external/npm` path.