Fix deployment of KeepRegistryStub
When a KeepRegistry deployed contract is not found for local network we
wan to deploy a stub. The script can also be used by external projects
to deploy threshold network contract dependencies. When we used
`keepDeployer` the external projects deployment was failing as it
required the account to be defined also in the other project's hardhat
configuration file. To fix that we use `deployer` named account to
deploy the stub, as it will for sure be defined in other projects.

To call the `approveOperatorContract` we use `keepRegistryKeeper` which
is defined by default as account `0` and it matches the `deployer`.