Staking requestAuthorizationDecrease function contract update
The application may revert authorization decrease request, if there is
already one pending for the given application.

This is entirely up to the application if overwriting pending request at
the given moment is fine or not.

This is important for TBTC and Random Beacon because allowing to
"decrease authorization decrease request" at any moment may be used for
pool manipulation and becoming a free-rider.