Removed hardhat-local-networks-config plugin
After an update in `deepmerge` library, `hardhat-local-networks-config`
stopped working and that lead to our system tests to fail with a
mysterious `TypeError: config.chains.has is not a function` error.

Until the problem is not fixed in `hardhat-local-networks-config`, the
plugin is removed. It was here for convenience and all tests are passing
without it.

See https://github.com/NomicFoundation/hardhat/issues/2630