Update `NuCypherTokenStakingMock` contract
Expose the `stakerInfo` function that returns the `staker` info to test
the T dapp locally and on Ropsten network. In the original impl of this
contract the `StakerStruct` is stored in the `stakerInfo` map not
`stakers`[1]. Renaming the `stakers` to `stakerInfo` in stub contract
requires a lot of changes in tests so here we expose the `stakeInfo`
function to match the stub contract ABI with the original one.

Ref:
- [1]: https://github.com/nucypher/nucypher-contracts/blob/main/contracts/contracts/StakingEscrow.sol#L151