Downgrade `hardhat-deploy` dependency back to `^0.9.27`
Previously dependency got upgraded to `0.11.15` version because the `0.9.29`
version was not working correctly with the `@nomiclabs/hardhat-etherscan`
dependency in the `3.1.0` version. Since then we have downgraded the
`@nomiclabs/hardhat-etherscan` to `3.0.4` version (due to some other issue),
which seems to not clash with both `hardhat-deploy` in `0.9.29` and in `0.11.15`
version. And as we noticed that `hardhat-deploy@0.11.15` makes our unit tests
for `TokenStaking` fail (errors like `AssertionError: expected [ …(2), …(2) ] to
deeply equal [ …(2) ]`), we're now choosing to use `hardhat-deploy` in `^0.9.29`
version.