Use floating pragma for IApplication and IStaking interfaces
Pragma statements are fine to float when a contract is intended for
consumption by contracts from another project. This is the case for
IApplication and IStaking that are imported from tBTC v2 contracts.
If we do not use the floating pragma, all tBTC v2 contracts would need
to stay on 0.8.9 and now use the more recent Solidity versions.