Configure Etherscan API key for use with @nomiclabs/hardhat-etherscan
In PR #126 (https://github.com/threshold-network/solidity-contracts/pull/126)
we've upgraded the `hardhat-deploy` plugin to latest version (^0.11.15). As part
of this change, the way of storing Etherscan API key has changed - to be
different from how the key is stored for @nomiclabs/hardhat-etherscan plugin
use. When we merged those changes from `main` to the current feature branch, we
no longer had a correct format of the Etherscan API key config to use it with
the `@nomiclabs/hardhat-etherscan` plugin. As we don't want to verify contracts
using `hardhat-deploy` plugin anymore, we need to adjust the config to fit
`@nomiclabs/hardhat-etherscan`'s syntax.