Include `./contracts/test` folder in the published packages
Back in a day we decided to exclude the `./contracts/test` and `./test`
directories using `"!**/test/"` expression. This was done to avoid confusions, as
at that point we had thought the files are used only internally in the package.
Since we started using more features of hardhat it turned out that the deployment
scripts and artifacts are no longer used just internally and we need
`./contracts/test` files in other projects using
`@threshold-network/solidity-contracts` as a dependency. For some reason
(probably a bug in Yarn or Node.js) the `"!**/test/"` expression wasn't
excluding the `./contracts/test` when preparing a package, so there was no
problem reported when using those files in other projects. But recently we've
upgraded the version of Node.js (`v14` -> `v18`) used to build the
`@threshold-network/solidity-contracts` package, and the other projects started
to complain about some missing files from `./contracts/test`. It seems that the
newer version of Node.js interprets the wildcards/negation differently
and the ./contracts/test` files were no longer being included in the
package.

So now we're restoring the earlier behavior by removing the `"!**/test/",`
expression from `files`. This way we will still exclude the `./test`, but will
include `./contracts/test` in the package.