Set Node.js version to `18.15.0`
We want to prevent the problem with generation of artifacts during execution of
the `hardhat compile` command (part of `yarn deploy`). We saw in other project
that sometimes command was not not generating expected artifacts. The problem is
caused by the process silently quitting, which is related to the used version of
Node (as descibed in https://github.com/NomicFoundation/hardhat/issues/3877).
We've confirmed the problem is reproducible on `v18.16.0`, now we're downgrading
to `v18.15.0` to fix the issue.