Set Node.js version to 18.15.0
We want to fix problem with the `yarn build` (`hardhat compile`) command
sometimes not generating expected artifacts. The problem is caused by the
process silently quitting, which is related to the used version of Node (as
descibed in NomicFoundation/hardhat#3877). We've confirmed the problem is
reproducible on `v18.16.0`, now we're downgrading to `v18.15.0` to fix the
issue.